Move new test helpers to spec/helpers.js
